Move most App.svelte code into a Tutorial component so it can be generalized for other tools
